refactor(server-rust): defer live-file reclaim; consolidate tier read path into tier.rs
Three related changes to the hot/cold tiering path.

1. Defer live-file reclaim (remove the hole-punch). seal_loop no longer
   `fallocate`-punches a sealed segment's region in the live data file.
   Hole-punching a shared file races with the engines' in-flight lazy reads
   (sendfile / Body::FileRange) into the just-sealed tail, which would read
   zeros from the freed blocks. The live file is now append-only (no
   mutation after write), so that race is gone. Sealed bytes are still served
   from the chunk file / remote object; the live file just retains a
   redundant copy of its sealed prefix (extra disk). Safe live-file reclaim
   (read/punch coordination via epoch/refcount, or compaction) is a planned
   follow-up. punch_hole is removed.

2. Consolidate the three range-walkers into one resolver. range_is_all_local
   and collect_segments are removed; resolve_range is now the single read
   resolver, and into_local_segments(slices) decides the body: all-Local →
   zero-copy Body::FileRange (including sealed chunk files, which previously
   forced the buffered path), any Remote → streamed Body::Channel. The read
   path resolves once instead of walking the manifest twice (range check then
   resolve), and the channel path no longer re-resolves inside its task.

3. Relocate all tiering logic to tier.rs. The lifecycle (seal/offload/GC/boot
   reconcile) moves to a second inherent `impl Store` block in tier.rs, and
   the resolver (resolve_range / resolve_sealed / into_local_segments /
   ResolvedSlice) moves there too; store.rs re-exports the resolver so callers
   keep a single `store::` facade. store.rs shrinks ~560 lines.

The sealed-Local regression test is repurposed to offload-failure resilience
(a sealed segment whose upload fails stays Local and reads back byte-identical
from its chunk file).

Verified: cargo test (default + tier, 23 pass), clippy clean (default +
tier), conformance 332/332 (tier off and tier local, 8 sealed objects), and a
uring+tier Docker smoke (append → seal → offload → cold whole-stream read
byte-identical).
